"Just good tools for thinking about anything"
I tend to be skeptical as soon as I hear or read the word "just."  This word often is used to pronounce an innocence that is entirely unfounded, but strongly desired.  Imagine Descartes saying, as he did, that his tools for the direction of thought were universal and good for any sort of problem.  Were they?  If so, we wouldn't have needed more, or new directions in thought that were subsequently required would have derived from his directions.  I strongly doubt there is an identity -- a something -- called "mathematics" that has the properties that are claimed for it here.